+++ pydriller/__init__.py
@@ -1,5 +1,5 @@
 # pylint: disable=C0111
 from .domain.commit import Commit, Modification, ModificationType # noqa
-from .repository_mining import Repository, Git    # noqa
+from .repository import Repository, Git    # noqa
 
 __version__ = "1.15.2"
